Cleaning up workflows

+++ aoclib/python/tests/test_helpers.py
@@ -1,29 +1,32 @@
-import os
+import typing
 import pathlib
-from unittest.mock import patch
+import contextlib
 
 import pytest
 from aoclib.helpers import load_input, splitlines, zip_longest_repeating
 
 
 class TestAoclib:
-    def test_load_input(self, tmp_path: pathlib.Path) -> None:
+    @pytest.mark.parametrize(
+        ("exists", "cm"),
+        [
+            (False, pytest.raises(FileNotFoundError, match="No such file or directory")),
+            (True, contextlib.nullcontext()),
+        ],
+        ids=("missing", "exists"),
+    )
+    def test_load_input(self, exists: bool, cm: typing.ContextManager, tmp_path: pathlib.Path) -> None:
+        pytest.MonkeyPatch().chdir(tmp_path)
         input_string = b"hello\nline\ntwo\nfour"
         expected = "hello\nline\ntwo\nfour"
 
-        file = tmp_path.joinpath("inputs", "2020", "15.txt")
-        file.parent.mkdir(parents=True, exist_ok=True)
+        if exists:
+            file = tmp_path.joinpath("input")
+            file.parent.mkdir(parents=True, exist_ok=True)
+            file.write_bytes(input_string)
 
-        file.write_bytes(input_string)
-
-        with patch.dict(os.environ, {"AOC_ROOT_DIRECTORY": str(tmp_path)}):
-            assert load_input("2020", "15") == expected
-
-            with pytest.raises(FileNotFoundError, match="No such file or directory"):
-                expected = load_input("2020", "16")
-
-        with pytest.raises(SystemExit, match=r"Could not determine the proper \$AOC_ROOT_DIRECTORY"):
-            expected = load_input("2020", "15")
+        with cm:
+            assert load_input() == expected
 
     def test_zip_longest_repeating(self) -> None:
         assert list(zip_longest_repeating(["a"], ["b", "c"], ["d", "r", "f"])) == [
